requirements.txt
Flask
openai
google-api-python-client
python-dotenv
